fix(cli): reject stale modified submissions at consent boundary
